+++ packages/react-native-sdk/src/hooks/push/useIosCallKeepEventsSetupEffect.ts
@@ -2,15 +2,22 @@ import { useEffect } from 'react';
 import {
   voipCallkeepCallOnForegroundMap$,
   voipPushNotificationCallCId$,
-} from '../../utils/push/rxSubjects';
+} from '../../utils/push/internal/rxSubjects';
 import { RxUtils } from '@stream-io/video-client';
-import {
-  iosCallkeepAcceptCall,
-  iosCallkeepRejectCall,
-} from '../../utils/push/ios';
 import { getCallKeepLib } from '../../utils/push/libs';
-import { StreamVideoRN } from '../../utils';
+import { StreamVideoRN } from '../../utils/StreamVideoRN';
+import type { StreamVideoConfig } from '../../utils/StreamVideoRN/types';
+import {
+  clearPushWSEventSubscriptions,
+  processCallFromPushInBackground,
+} from '../../utils/push/internal/utils';
+import {
+  pushAcceptedIncomingCallCId$,
+  voipCallkeepAcceptedCallOnNativeDialerMap$,
+} from '../../utils/push/internal/rxSubjects';
 import { Platform } from 'react-native';
+
+type PushConfig = NonNullable<StreamVideoConfig['push']>;
 
 /**
  * This hook is used to listen to callkeep events and do the necessary actions
@@ -61,4 +68,53 @@ export const useIosCallKeepEventsSetupEffect = () => {
       removeDisplayIncomingCall();
     };
   }, []);
+};
+
+const iosCallkeepAcceptCall = (
+  call_cid: string | undefined,
+  callUUIDFromCallkeep: string
+) => {
+  if (!shouldProcessCallFromCallkeep(call_cid, callUUIDFromCallkeep)) {
+    return;
+  }
+  clearPushWSEventSubscriptions();
+  // to call end callkeep later if ended in app and not through callkeep
+  voipCallkeepAcceptedCallOnNativeDialerMap$.next({
+    uuid: callUUIDFromCallkeep,
+    cid: call_cid,
+  });
+  // to process the call in the app
+  pushAcceptedIncomingCallCId$.next(call_cid);
+  // no need to keep these references anymore
+  voipCallkeepCallOnForegroundMap$.next(undefined);
+};
+
+const iosCallkeepRejectCall = async (
+  call_cid: string | undefined,
+  callUUIDFromCallkeep: string,
+  pushConfig: PushConfig
+) => {
+  if (!shouldProcessCallFromCallkeep(call_cid, callUUIDFromCallkeep)) {
+    return;
+  }
+  clearPushWSEventSubscriptions();
+  // no need to keep these references anymore
+  voipCallkeepAcceptedCallOnNativeDialerMap$.next(undefined);
+  voipCallkeepCallOnForegroundMap$.next(undefined);
+  voipPushNotificationCallCId$.next(undefined);
+  await processCallFromPushInBackground(pushConfig, call_cid, 'decline');
+};
+
+/**
+ * Helper function to determine if the answer/end call event from callkeep must be processed
+ * Just checks if we have a valid call_cid and acts as a type guard for call_cid
+ */
+const shouldProcessCallFromCallkeep = (
+  call_cid: string | undefined,
+  callUUIDFromCallkeep: string
+): call_cid is string => {
+  if (!call_cid || !callUUIDFromCallkeep) {
+    return false;
+  }
+  return true;
 };
