+++ packages/react-native-sdk/src/hooks/push/useProcessPushNonRingingCallEffect.ts
@@ -1,4 +1,4 @@
-import { pushNonRingingCallData$ } from '../../utils/push/rxSubjects';
+import { pushNonRingingCallData$ } from '../../utils/push/internal/rxSubjects';
 import { useEffect } from 'react';
 import { StreamVideoRN } from '../../utils';
 import {
@@ -6,7 +6,7 @@ import {
   useStreamVideoClient,
 } from '@stream-io/video-react-bindings';
 import { filter } from 'rxjs/operators';
-import { processNonIncomingCallFromPush } from '../../utils/push/utils';
+import { processNonIncomingCallFromPush } from '../../utils/push/internal/utils';
 
 /**
  * This hook is used to process the non ringing call data via push notifications using the relevant rxjs subject
