+++ packages/react-native-sdk/src/utils/push/android.ts
@@ -3,7 +3,6 @@ import notifee, {
   Event,
   AndroidCategory,
 } from '@notifee/react-native';
-import { FirebaseMessagingTypes } from '@react-native-firebase/messaging';
 import {
   Call,
   RxUtils,
@@ -17,9 +16,8 @@ import type {
 } from '../StreamVideoRN/types';
 import {
   getFirebaseMessagingLib,
-  getFirebaseMessagingLibNoThrow,
   getExpoNotificationsLib,
-  getExpoTaskManagerLib,
+  FirebaseMessagingTypes,
 } from './libs';
 import {
   pushAcceptedIncomingCallCId$,
@@ -28,99 +26,21 @@ import {
   pushNonRingingCallData$,
   pushUnsubscriptionCallbacks$,
   pushAndroidBackgroundDeliveredIncomingCallCId$,
-} from './rxSubjects';
+} from './internal/rxSubjects';
 import {
   canAddPushWSSubscriptionsRef,
   clearPushWSEventSubscriptions,
   processCallFromPushInBackground,
   shouldCallBeEnded,
-} from './utils';
+} from './internal/utils';
 import { setPushLogoutCallback } from '../internal/pushLogoutCallback';
 import { getAndroidDefaultRingtoneUrl } from '../getAndroidDefaultRingtoneUrl';
+import { StreamVideoRN } from '../StreamVideoRN';
 
 const ACCEPT_CALL_ACTION_ID = 'accept';
 const DECLINE_CALL_ACTION_ID = 'decline';
 
 type PushConfig = NonNullable<StreamVideoConfig['push']>;
-
-/** Setup Firebase push message handler **/
-export function setupFirebaseHandlerAndroid(pushConfig: PushConfig) {
-  if (Platform.OS !== 'android') {
-    return;
-  }
-  if (pushConfig.isExpo) {
-    const messaging = getFirebaseMessagingLibNoThrow(true);
-    if (messaging) {
-      // handles on app killed state in expo, expo-notifications cannot handle that
-      messaging().setBackgroundMessageHandler(
-        async (msg) =>
-          await firebaseMessagingOnMessageHandler(msg.data, pushConfig)
-      );
-      messaging().onMessage((msg) =>
-        firebaseMessagingOnMessageHandler(msg.data, pushConfig)
-      ); // this is to listen to foreground messages, which we dont need for now
-    } else {
-      const Notifications = getExpoNotificationsLib();
-      const TaskManager = getExpoTaskManagerLib();
-      const BACKGROUND_NOTIFICATION_TASK =
-        'STREAM-VIDEO-SDK-INTERNAL-BACKGROUND-NOTIFICATION-TASK';
-
-      TaskManager.defineTask(
-        BACKGROUND_NOTIFICATION_TASK,
-        ({ data, error }) => {
-          if (error) {
-            return;
-          }
-          // @ts-ignore
-          const dataToProcess = data.notification?.data;
-          firebaseMessagingOnMessageHandler(dataToProcess, pushConfig);
-        }
-      );
-      // background handler (does not handle on app killed state)
-      Notifications.registerTaskAsync(BACKGROUND_NOTIFICATION_TASK);
-      // foreground handler
-      Notifications.setNotificationHandler({
-        handleNotification: async (notification) => {
-          // @ts-ignore
-          const trigger = notification?.request?.trigger;
-          if (trigger.type === 'push') {
-            const data = trigger?.remoteMessage?.data;
-            if (data?.sender === 'stream.video') {
-              await firebaseMessagingOnMessageHandler(data, pushConfig);
-              return {
-                shouldShowAlert: false,
-                shouldPlaySound: false,
-                shouldSetBadge: false,
-              };
-            }
-          }
-          return {
-            shouldShowAlert: true,
-            shouldPlaySound: false,
-            shouldSetBadge: false,
-          };
-        },
-      });
-    }
-  } else {
-    const messaging = getFirebaseMessagingLib();
-    messaging().setBackgroundMessageHandler(
-      async (msg) =>
-        await firebaseMessagingOnMessageHandler(msg.data, pushConfig)
-    );
-    messaging().onMessage((msg) =>
-      firebaseMessagingOnMessageHandler(msg.data, pushConfig)
-    ); // this is to listen to foreground messages, which we dont need for now
-  }
-
-  // the notification tap handlers are always registered with notifee for both expo and non-expo in android
-  notifee.onBackgroundEvent(async (event) => {
-    await onNotifeeEvent(event, pushConfig, true);
-  });
-  notifee.onForegroundEvent((event) => {
-    onNotifeeEvent(event, pushConfig, false);
-  });
-}
 
 /** Send token to stream, create notification channel,  */
 export async function initAndroidPushToken(
@@ -166,10 +86,14 @@ export async function initAndroidPushToken(
   }
 }
 
+/**
+ * Creates notification from the push message data.
+ * For Ringing and Non-Ringing calls.
+ */
 export const firebaseMessagingOnMessageHandler = async (
-  data: FirebaseMessagingTypes.RemoteMessage['data'],
-  pushConfig: PushConfig
+  message: FirebaseMessagingTypes.RemoteMessage
 ) => {
+  const data = message.data;
   /* Example data from firebase
     "message": {
         "data": {
@@ -185,7 +109,8 @@ export const firebaseMessagingOnMessageHandler = async (
         // other stuff
     }
   */
-  if (!data || data.sender !== 'stream.video') {
+  const pushConfig = StreamVideoRN.getConfig().push;
+  if (!pushConfig || !data || data.sender !== 'stream.video') {
     return;
   }
 
@@ -354,16 +279,24 @@ export const firebaseMessagingOnMessageHandler = async (
   }
 };
 
-const onNotifeeEvent = async (
-  event: Event,
-  pushConfig: PushConfig,
-  isBackground: boolean
-) => {
+export const onAndroidNotifeeEvent = async ({
+  event,
+  isBackground,
+}: {
+  event: Event;
+  isBackground: boolean;
+}) => {
   const { type, detail } = event;
   const { notification, pressAction } = detail;
   const notificationId = notification?.id;
   const data = notification?.data;
-  if (!data || !notificationId || data.sender !== 'stream.video') {
+  const pushConfig = StreamVideoRN.getConfig().push;
+  if (
+    !pushConfig ||
+    !data ||
+    !notificationId ||
+    data.sender !== 'stream.video'
+  ) {
     return;
   }
 
@@ -413,7 +346,8 @@ const onNotifeeEvent = async (
     }
   } else {
     if (type === EventType.PRESS) {
-      pushTappedIncomingCallCId$.next(call_cid);
+      // NOTE: the firebase handler sets pushNonRingingCallData$ so its not necessary to set here
+      // here we handle only the press listener, so we call the callback
       pushConfig.onTapNonRingingCallNotification?.(
         call_cid,
         data.type as NonRingingPushEvent
