+++ packages/react-native-sdk/src/utils/push/ios.ts
@@ -1,36 +1,22 @@
 import { Platform } from 'react-native';
-import type {
-  NonRingingPushEvent,
-  StreamVideoConfig,
-} from '../StreamVideoRN/types';
+import type { StreamVideoConfig } from '../StreamVideoRN/types';
+import { pushNonRingingCallData$ } from './internal/rxSubjects';
 import {
-  pushAcceptedIncomingCallCId$,
-  voipPushNotificationCallCId$,
-  voipCallkeepCallOnForegroundMap$,
-  voipCallkeepAcceptedCallOnNativeDialerMap$,
-  pushNonRingingCallData$,
-} from './rxSubjects';
-import {
-  clearPushWSEventSubscriptions,
-  processCallFromPushInBackground,
-} from './utils';
-import { getExpoNotificationsLib, getPushNotificationIosLib } from './libs';
+  ExpoNotification,
+  getExpoNotificationsLib,
+  getPushNotificationIosLib,
+  PushNotificationiOSType,
+} from './libs';
 import { StreamVideoClient, getLogger } from '@stream-io/video-client';
 import { setPushLogoutCallback } from '../internal/pushLogoutCallback';
-import notifee, { EventType } from '@notifee/react-native';
+import { EventType, Event } from '@notifee/react-native';
+import { StreamVideoRN } from '../StreamVideoRN';
+import { StreamPushPayload } from './utils';
 
 type PushConfig = NonNullable<StreamVideoConfig['push']>;
 
-type StreamPayload =
-  | {
-      call_cid: string;
-      type: 'call.ring' | NonRingingPushEvent;
-      sender: string;
-    }
-  | undefined;
-
 function processNonRingingNotificationStreamPayload(
-  streamPayload: StreamPayload
+  streamPayload: StreamPushPayload
 ) {
   if (
     streamPayload?.sender === 'stream.video' &&
@@ -43,85 +29,55 @@ function processNonRingingNotificationStreamPayload(
   }
 }
 
-export const iosCallkeepAcceptCall = (
-  call_cid: string | undefined,
-  callUUIDFromCallkeep: string
-) => {
-  if (!shouldProcessCallFromCallkeep(call_cid, callUUIDFromCallkeep)) {
-    return;
-  }
-  clearPushWSEventSubscriptions();
-  // to call end callkeep later if ended in app and not through callkeep
-  voipCallkeepAcceptedCallOnNativeDialerMap$.next({
-    uuid: callUUIDFromCallkeep,
-    cid: call_cid,
-  });
-  // to process the call in the app
-  pushAcceptedIncomingCallCId$.next(call_cid);
-  // no need to keep these references anymore
-  voipCallkeepCallOnForegroundMap$.next(undefined);
-};
-
-export const iosCallkeepRejectCall = async (
-  call_cid: string | undefined,
-  callUUIDFromCallkeep: string,
-  pushConfig: PushConfig
-) => {
-  if (!shouldProcessCallFromCallkeep(call_cid, callUUIDFromCallkeep)) {
-    return;
-  }
-  clearPushWSEventSubscriptions();
-  // no need to keep these references anymore
-  voipCallkeepAcceptedCallOnNativeDialerMap$.next(undefined);
-  voipCallkeepCallOnForegroundMap$.next(undefined);
-  voipPushNotificationCallCId$.next(undefined);
-  await processCallFromPushInBackground(pushConfig, call_cid, 'decline');
-};
-
-/**
- * Helper function to determine if the answer/end call event from callkeep must be processed
- * Just checks if we have a valid call_cid and acts as a type guard for call_cid
- */
-const shouldProcessCallFromCallkeep = (
-  call_cid: string | undefined,
-  callUUIDFromCallkeep: string
-): call_cid is string => {
-  if (!call_cid || !callUUIDFromCallkeep) {
-    return false;
-  }
-  return true;
-};
-
-export const setupRemoteNotificationsHandleriOS = (pushConfig: PushConfig) => {
-  if (Platform.OS !== 'ios') {
-    return;
-  }
-  notifee.onForegroundEvent(({ type, detail }) => {
-    if (type === EventType.PRESS) {
-      const streamPayload = detail.notification?.data?.stream as
-        | StreamPayload
-        | undefined;
-      const result = processNonRingingNotificationStreamPayload(streamPayload);
-      if (result) {
-        pushConfig.onTapNonRingingCallNotification?.(result.cid, result.type);
-      }
+export const oniOSExpoNotificationEvent = (event: ExpoNotification) => {
+  const pushConfig = StreamVideoRN.getConfig().push;
+  if (pushConfig) {
+    if (event.request.trigger.type === 'push') {
+      const streamPayload = event.request.trigger.payload
+        ?.stream as StreamPushPayload;
+      processNonRingingNotificationStreamPayload(streamPayload);
     }
-  });
-  if (pushConfig.isExpo) {
-    const Notifications = getExpoNotificationsLib();
-
-    // foreground handler (just to show the notifications on foreground)
-    Notifications.setNotificationHandler({
-      handleNotification: async () => {
-        return {
-          shouldShowAlert: true,
-          shouldPlaySound: true,
-          shouldSetBadge: false,
-        };
-      },
-    });
   }
 };
+
+export const oniOSNotifeeEvent = ({
+  event,
+}: {
+  event: Event;
+  isBackground: boolean;
+}) => {
+  const pushConfig = StreamVideoRN.getConfig().push;
+  const { type, detail } = event;
+  if (pushConfig && type === EventType.PRESS) {
+    const streamPayload = detail.notification?.data?.stream as
+      | StreamPushPayload
+      | undefined;
+    const result = processNonRingingNotificationStreamPayload(streamPayload);
+    if (result) {
+      pushConfig.onTapNonRingingCallNotification?.(result.cid, result.type);
+    }
+  }
+};
+
+export function onPushNotificationiOSStreamVideoEvent(
+  notification: PushNotificationiOSType
+) {
+  const pushNotificationIosLib = getPushNotificationIosLib();
+  const data = notification.getData();
+  const isClicked = data.userInteraction === 1;
+  const pushConfig = StreamVideoRN.getConfig().push;
+  if (!isClicked || !pushConfig) {
+    notification.finish(pushNotificationIosLib.FetchResult.NoData);
+    return;
+  }
+  const streamPayload = data?.stream as StreamPushPayload;
+  // listen to foreground notifications
+  const result = processNonRingingNotificationStreamPayload(streamPayload);
+  if (result) {
+    pushConfig.onTapNonRingingCallNotification?.(result.cid, result.type);
+  }
+  notification.finish(pushNotificationIosLib.FetchResult.NoData);
+}
 
 /** Send token to stream */
 export async function initIosNonVoipToken(
@@ -158,34 +114,16 @@ export async function initIosNonVoipToken(
         setDeviceToken(devicePushToken.data);
       }
     );
-    const subscriptionForReceive =
-      expoNotificationsLib.addNotificationReceivedListener((event) => {
-        // listen to foreground notifications
-        if (event.request.trigger.type === 'push') {
-          const streamPayload = event.request.trigger.payload
-            ?.stream as StreamPayload;
-          processNonRingingNotificationStreamPayload(streamPayload);
-        }
-      });
     setUnsubscribeListener(() => {
       subscription.remove();
-      subscriptionForReceive.remove();
     });
   } else {
     const pushNotificationIosLib = getPushNotificationIosLib();
     pushNotificationIosLib.addEventListener('register', (token) => {
       setDeviceToken(token);
     });
-    pushNotificationIosLib.addEventListener('notification', (notification) => {
-      const data = notification.getData();
-      const streamPayload = data?.stream as StreamPayload;
-      // listen to foreground notifications
-      processNonRingingNotificationStreamPayload(streamPayload);
-      notification.finish(pushNotificationIosLib.FetchResult.NoData);
-    });
     setUnsubscribeListener(() => {
       pushNotificationIosLib.removeEventListener('register');
-      pushNotificationIosLib.removeEventListener('notification');
     });
   }
 }
