+++ packages/react-native-sdk/src/utils/push/libs/expoNotifications.ts
@@ -1,5 +1,9 @@
 export type ExpoNotificationsLib = typeof import('expo-notifications');
 
+import type { Notification } from 'expo-notifications';
+
+export type ExpoNotification = Notification;
+
 let expoNotificationsLib: ExpoNotificationsLib | undefined;
 
 try {
